REST_API_Vulnerabilities_Python/web_app/requirements.txt
Django >= 4.0
djangorestframework>=3.12
psycopg2-binary